requirements.txt
streamlit
supabase
pandas
plotly
numpy
python-dotenv
groq